+++ ttt/class/app/app_arguments.h
@@ -99,6 +99,9 @@ class AppArguments {
   std::string ResolveTime(int index, bool allow_negative = false) const;
   // Remove possible prefix from comment ("c=" or "comment=")
   std::string ResolveComment(int index) const;
+
+  bool ResolveYesNo(int index) const;
+
   // Validate command-name from argument at given index
   AppCommand::Commands ResolveCommandName(int index) const;